bug fixes, calibration (heat up) time changed for gas sensor, analog value is used now

+++ printerHousingCtrl/printerHousingCtrl/gasSensor.h
@@ -12,7 +12,7 @@
 #include "fireExtinguisher.h"
 
 typedef struct{
-	float gasSensAnalog;
+	uint16_t gasSensAnalog;
 	bool gasSensDig;
 	bool fireDetected;
 	uint32_t fireDetectedTime;
